requirements.txt

selenium==3.141.0
appium-python-client==1.0.2
opencv-python==4.5.3
pytesseract==0.3.8
pandas==1.3.3
